Added calls for server-fs

+++ src/kernel/vfs/file_table.c
@@ -1,7 +1,7 @@
 #include "vfs.h"
 #include "vfs_alloc.h"
 
-FtEntry_t* ft_add_entry(ServiceHandle_t handle, long file_id, int attributes)
+FtEntry_t* ft_add_entry(ServiceHandle_t handle, u64 file_id, u64 attributes)
 {
     FtEntry_t* ret = (FtEntry_t*) vfs_malloc(sizeof(FtEntry_t));
     ret->handle = handle;
